Move ErrorResponse into HttpPipelineExtensions
